+++ src/Control/Monad/ST/Unsafe/Compat.hs
@@ -6,4 +6,8 @@ module Control.Monad.ST.Unsafe.Compat (
 , unsafeSTToIO
 ) where
 
+#if MIN_VERSION_base(4,6,0)
+import Control.Monad.ST.Unsafe (unsafeInterleaveST, unsafeIOToST, unsafeSTToIO)
+#else
 import Control.Monad.ST (unsafeInterleaveST, unsafeIOToST, unsafeSTToIO)
+#endif