+++ src/Foreign/Marshal/Unsafe/Compat.hs
@@ -4,4 +4,8 @@ module Foreign.Marshal.Unsafe.Compat (
   unsafeLocalState
 ) where
 
+#if MIN_VERSION_base(4,6,0)
+import Foreign.Marshal.Unsafe (unsafeLocalState)
+#else
 import Foreign.Marshal (unsafeLocalState)
+#endif